Modify: refactor component folder
