+++ .eslintignore
@@ -1,0 +1,4 @@
+node_modules/
+package.json
+client-dist/
+server-dist/
