+++ client/components/Event/EventInfo.jsx
