+++ client/components/Event/PopularCenter.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import Navbar from './Navbar';
+import Navbar from '../common/Navbar';
 
 const PopularEvents = ({
   events,
